Started working on cloning entities.

+++ src/main/java/com/dietator/diet/controller/MealController.java
@@ -15,7 +15,7 @@ public class MealController {
     private final MealService mealService;
 
     @GetMapping("/meals/{id}")
-    public Meal getMeal(@PathVariable int id) {
+    public Meal getMeal(@PathVariable Long id) {
         return mealService.getMealById(id);
     }
 
@@ -35,7 +35,7 @@ public class MealController {
     }
 
     @DeleteMapping("/meals/{id}")
-    public void deleteMeal(@PathVariable int id) {
+    public void deleteMeal(@PathVariable Long id) {
         mealService.deleteMeal(id);
     }
 }
